fix EventStoreCatchUpSubscription interfaces
