+++ src/EventStoreCatchUpSubscription.php
@@ -1,0 +1,32 @@
+<?php
+
+/**
+ * This file is part of prooph/event-store.
+ * (c) 2014-2018 prooph software GmbH <contact@prooph.de>
+ * (c) 2015-2018 Sascha-Oliver Prolic <saschaprolic@googlemail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Prooph\EventStore;
+
+use Amp\Promise;
+use Throwable;
+
+interface EventStoreCatchUpSubscription
+{
+    public function isSubscribedToAll(): bool;
+
+    public function streamId(): string;
+
+    public function subscriptionName(): string;
+
+    public function start(): void;
+
+    public function stop(): void;
+
+    public function dropSubscription(SubscriptionDropReason $reason, ?Throwable $error): void;
+}
