--- src/Util/EventStoreStreamCatchUpSubscription.php
@@ -1,19 +1,0 @@
-<?php
-
-/**
- * This file is part of prooph/event-store.
- * (c) 2014-2018 prooph software GmbH <contact@prooph.de>
- * (c) 2015-2018 Sascha-Oliver Prolic <saschaprolic@googlemail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Prooph\EventStore;
-
-interface EventStoreStreamCatchUpSubscription extends EventStoreCatchUpSubscription
-{
-    public function lastProcessedEventNumber(): int;
-}
